requirements.txt
torch
torchvision
pandas 
numpy 
scikit-learn
tensorboard
matplotlib
streamlit
optuna
optuna-integration[tfkeras]
plotly
Pillow
